Validation fixes

+++ src/Repository/RepoRepository.php
@@ -5,6 +5,7 @@ namespace App\Repository;
 use App\Entity\Org;
 use App\Entity\Repo;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\DBAL\Exception;
 use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\Persistence\ManagerRegistry;
 
@@ -28,20 +29,20 @@ class RepoRepository extends ServiceEntityRepository
         return $qb->getQuery()->getResult();
     }
 
+    /**
+     * @throws Exception
+     */
     public function deleteOfType($provider, $orgName)
     {
         $conn = $this->getEntityManager()->getConnection();
 
         $sql = '
-            DELETE FROM repo
-            USING org
-            WHERE repo.provider = :provider
-            AND org.name = :orgName
+            DELETE FROM repo AS r USING org AS o WHERE r.org_id = o.id AND o.name = :orgName AND r.provider = :provider
             ';
         $stmt = $conn->prepare($sql);
-        $resultSet = $stmt->executeQuery(['provider' => $provider, 'orgName' => $orgName]);
 
-        return $resultSet->fetchAllAssociative();
+        $stmt->executeQuery(['provider' => $provider,'orgName' => $orgName]);
+
 
     }
 
@@ -49,7 +50,7 @@ class RepoRepository extends ServiceEntityRepository
     {
         return $this->createQueryBuilder('r')
             
-            ->innerJoin(Org::class, 'o', \Doctrine\ORM\Query\Expr\Join::WITH, 'r.org = o.id')
+            ->innerJoin(Org::class, 'o', Join::WITH, 'r.org = o.id')
             ->where('r INSTANCE OF :provider')
             ->andWhere('o.name = :orgName')
             ->setParameter('provider', $provider)
